Move dep tree out of get-deps-tree.js
