+++ script/dep-map-json.js
@@ -1,0 +1,3 @@
+'use strict'
+const {dep, tryReadJSON} = require('../lib/common-vars.js')
+module.exports = tryReadJSON(require('path').join(dep, 'dependencies.json'))
