+++ script/get-deps-tree.js
@@ -1,30 +1,8 @@
 'use strict'
 
-const {join} = require('path')
-const {statSync, readdirSync} = require('fs')
 const merge = require('deepmerge')
-const jsDepTree = require('dependency-tree')
-const {projdir, lib, src, dep, tryReadJSON} = require('../lib/common-vars.js')
 const depYAML = require('./dep-map-yml.js')
-const depJSON = tryReadJSON(join(dep, 'dependencies.json'))
-const depNODE = {}
-
-; [lib, src, join(projdir, 'script')].forEach(genNodeDeps)
-
-function genNodeDeps (name) {
-  const stats = statSync(name)
-  if (stats.isFile()) {
-    if (/\.js$/.test(name)) {
-      const filter = string => !/node_modules/.test(string)
-      depNODE[name] = jsDepTree.toList({filename: name, directory: projdir, filter})
-    } else {
-      depNODE[name] = []
-    }
-  } else if (stats.isDirectory()) {
-    readdirSync(name).map(item => join(name, item)).forEach(genNodeDeps)
-  } else {
-    throw new Error(`Unreconizable file system entry: ${name}`)
-  }
-}
+const depJSON = require('./dep-map-json.js')
+const depNODE = require('./dep-map-deduced.js')
 
 module.exports = merge.all([depYAML, depJSON, depNODE])
